backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
chromadb>=0.5.0
sentence-transformers>=3.0.0
groq>=0.9.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.7.0
